refactor(shared): move GameMessage and friends from game/ to shared/ (no behavior change)
