+++ game/src/messages.rs
@@ -1,171 +1,75 @@
-use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
+//! Game-side message helpers.
+//!
+//! The schema types (`GameMessage`, `MessageKind`, `MessageSource`) live in
+//! the `shared` crate so the `web` and `api` crates can use them without
+//! pulling in `game`. This module re-exports them and adds the
+//! game-only conveniences:
+//!
+//! - Constructors that take a structured `GameEvent` payload from
+//!   [`crate::events`] (to be removed in Task 3 of the timeline-pr1
+//!   refactor).
+//! - Terrain-aware narrative helpers used by tribute/area logic.
+
+pub use shared::messages::{GameMessage, MessageKind, MessageSource};
+
 use uuid::Uuid;
 
 use crate::events::GameEvent;
 use crate::terrain::{BaseTerrain, Harshness, Visibility};
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(tag = "type", content = "value")]
-pub enum MessageSource {
-    #[serde(rename = "Game")]
-    Game(String), // Game identifier
-    #[serde(rename = "Area")]
-    Area(String), // Area name
-    #[serde(rename = "Tribute")]
-    Tribute(String), // Tribute identifier
+/// Decode a `GameMessage`'s structured event payload into a strong-typed
+/// [`GameEvent`]. Returns `None` if no payload is attached.
+///
+/// **Temporary**: the underlying `event` field is removed in Task 3 of the
+/// game-timeline-pr1 plan. Prefer not to add new callers.
+pub fn structured_event(message: &GameMessage) -> Option<Result<GameEvent, serde_json::Error>> {
+    message
+        .event
+        .as_deref()
+        .map(serde_json::from_str::<GameEvent>)
 }
 
-/// Typed category for a `GameMessage`. Initial set covers alliance lifecycle
-/// events; future categories (combat, area, sponsor) will be added as those
-/// emit sites get refactored. `None` on `GameMessage.kind` means the message
-/// has not yet been categorised.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub enum MessageKind {
-    AllianceFormed,
-    BetrayalTriggered,
-    TrustShockBreak,
+/// Create a new game message carrying a structured [`GameEvent`] payload.
+/// `content` is typically the rendered `Display` form of `event`, kept as
+/// a denormalised string so legacy consumers still work. Generates a
+/// fresh `event_id` so future dedup / hashing has a stable handle.
+///
+/// **Temporary**: the underlying `event` / `event_id` fields are removed
+/// in Task 3 of the game-timeline-pr1 plan.
+pub fn with_event(
+    source: MessageSource,
+    game_day: u32,
+    subject: String,
+    content: String,
+    event: GameEvent,
+) -> GameMessage {
+    let event_json =
+        serde_json::to_string(&event).expect("GameEvent serialization is infallible");
+    let mut msg = GameMessage::new(source, game_day, subject, content);
+    msg.event = Some(event_json);
+    msg.event_id = Some(Uuid::new_v4().to_string());
+    msg
 }
 
-/// Transcoding helpers for the structured event payload on
-/// [`GameMessage`]. The field is stored as a JSON-encoded `String` so it
-/// survives SurrealDB's bespoke (de)serializer (which collapses
-/// externally-tagged Rust enums to `{}` when bound into an `object`
-/// column); these helpers give callers strong-typed access without
-/// leaking the wire format.
-impl GameMessage {
-    /// Decode this message's structured event payload into a strong-typed
-    /// [`GameEvent`]. Returns `None` if no payload is attached.
-    pub fn structured_event(&self) -> Option<Result<GameEvent, serde_json::Error>> {
-        self.event.as_deref().map(serde_json::from_str::<GameEvent>)
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct GameMessage {
-    pub identifier: String,
-    pub source: MessageSource,
-    pub game_day: u32,
-    pub subject: String,
-    #[serde(with = "chrono::serde::ts_nanoseconds")]
-    pub timestamp: DateTime<Utc>,
-    pub content: String,
-    /// Optional typed category. `None` for legacy/uncategorised messages so
-    /// existing serialized rows hydrate cleanly. Skipped on serialize when
-    /// absent to keep JSON compact.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub kind: Option<MessageKind>,
-    /// Optional structured payload — the externally-tagged JSON form of
-    /// [`GameEvent`], serialized to a `String`. Stored as a string (not
-    /// a structured `serde_json::Value`) so it survives SurrealDB's
-    /// bespoke (de)serializer, which collapses externally-tagged Rust
-    /// enums to `{}` when bound into an `object` column. `None` for
-    /// legacy rows (and plain `log` calls that have no structured
-    /// event). Skipped on serialize when absent to keep JSON compact.
-    /// Use [`Self::structured_event`] to decode it back into a
-    /// strong-typed [`GameEvent`].
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub event: Option<String>,
-    /// Stable per-row identifier for the structured event payload, distinct
-    /// from `identifier` on the message itself. `Some` when `event` is
-    /// `Some`; `None` for legacy rows. Stored as a string-form UUID to
-    /// match the repo's existing on-the-wire UUID convention. Enables
-    /// future hashing / dedup of structured events without coupling to
-    /// message row identity.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub event_id: Option<String>,
-}
-
-impl GameMessage {
-    /// Create a new game message without a typed kind.
-    pub fn new(source: MessageSource, game_day: u32, subject: String, content: String) -> Self {
-        GameMessage {
-            identifier: Uuid::new_v4().to_string(),
-            source,
-            game_day,
-            subject,
-            timestamp: Utc::now(),
-            content,
-            kind: None,
-            event: None,
-            event_id: None,
-        }
-    }
-
-    /// Create a new game message with a typed `MessageKind`.
-    pub fn with_kind(
-        source: MessageSource,
-        game_day: u32,
-        subject: String,
-        content: String,
-        kind: MessageKind,
-    ) -> Self {
-        GameMessage {
-            identifier: Uuid::new_v4().to_string(),
-            source,
-            game_day,
-            subject,
-            timestamp: Utc::now(),
-            content,
-            kind: Some(kind),
-            event: None,
-            event_id: None,
-        }
-    }
-
-    /// Create a new game message carrying a structured [`GameEvent`] payload.
-    /// `content` is typically the rendered `Display` form of `event`, kept as
-    /// a denormalised string so legacy consumers still work. Generates a
-    /// fresh `event_id` so future dedup / hashing has a stable handle.
-    /// The event is JSON-encoded into a `String` for transport (see
-    /// [`Self::event`]); use [`Self::structured_event`] to decode it back.
-    pub fn with_event(
-        source: MessageSource,
-        game_day: u32,
-        subject: String,
-        content: String,
-        event: GameEvent,
-    ) -> Self {
-        let event_json =
-            serde_json::to_string(&event).expect("GameEvent serialization is infallible");
-        GameMessage {
-            identifier: Uuid::new_v4().to_string(),
-            source,
-            game_day,
-            subject,
-            timestamp: Utc::now(),
-            content,
-            kind: None,
-            event: Some(event_json),
-            event_id: Some(Uuid::new_v4().to_string()),
-        }
-    }
-
-    /// Create a new game message carrying both a typed [`MessageKind`] and a
-    /// structured [`GameEvent`] payload. See [`Self::with_event`] for the
-    /// `event_id` rationale and the JSON-on-the-wire note.
-    pub fn with_event_kind(
-        source: MessageSource,
-        game_day: u32,
-        subject: String,
-        content: String,
-        event: GameEvent,
-        kind: MessageKind,
-    ) -> Self {
-        let event_json =
-            serde_json::to_string(&event).expect("GameEvent serialization is infallible");
-        GameMessage {
-            identifier: Uuid::new_v4().to_string(),
-            source,
-            game_day,
-            subject,
-            timestamp: Utc::now(),
-            content,
-            kind: Some(kind),
-            event: Some(event_json),
-            event_id: Some(Uuid::new_v4().to_string()),
-        }
-    }
+/// Create a new game message carrying both a typed [`MessageKind`] and a
+/// structured [`GameEvent`] payload. See [`with_event`] for the
+/// `event_id` rationale and the JSON-on-the-wire note.
+///
+/// **Temporary**: removed in Task 3 of the game-timeline-pr1 plan.
+pub fn with_event_kind(
+    source: MessageSource,
+    game_day: u32,
+    subject: String,
+    content: String,
+    event: GameEvent,
+    kind: MessageKind,
+) -> GameMessage {
+    let event_json =
+        serde_json::to_string(&event).expect("GameEvent serialization is infallible");
+    let mut msg = GameMessage::with_kind(source, game_day, subject, content, kind);
+    msg.event = Some(event_json);
+    msg.event_id = Some(Uuid::new_v4().to_string());
+    msg
 }
 
 /// Generate terrain-aware movement narrative.
@@ -485,96 +389,14 @@ fn terrain_name(terrain: BaseTerrain) -> &'static str {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use chrono::TimeZone;
-    use serde_json::json;
 
     #[test]
-    fn message_kind_serde_roundtrip() {
-        for kind in [
-            MessageKind::AllianceFormed,
-            MessageKind::BetrayalTriggered,
-            MessageKind::TrustShockBreak,
-        ] {
-            let s = serde_json::to_string(&kind).expect("serialize MessageKind");
-            let back: MessageKind = serde_json::from_str(&s).expect("deserialize MessageKind");
-            assert_eq!(kind, back);
-        }
-    }
-
-    #[test]
-    fn game_message_kind_optional_field_default() {
-        // Legacy row: no `kind` field present.
-        let ts = Utc.timestamp_nanos(0);
-        let raw = json!({
-            "identifier": "abc",
-            "source": { "type": "Game", "value": "g1" },
-            "game_day": 1,
-            "subject": "game:g1",
-            "timestamp": ts.timestamp_nanos_opt().unwrap(),
-            "content": "hello",
-        });
-        let msg: GameMessage = serde_json::from_value(raw).expect("deserialize without kind");
-        assert!(msg.kind.is_none());
-    }
-
-    #[test]
-    fn game_message_with_kind_constructor_sets_kind() {
-        let msg = GameMessage::with_kind(
-            MessageSource::Game("g".into()),
-            2,
-            "game:g".into(),
-            "content".into(),
-            MessageKind::AllianceFormed,
-        );
-        assert_eq!(msg.kind, Some(MessageKind::AllianceFormed));
-    }
-
-    #[test]
-    fn game_message_default_constructor_has_no_kind() {
-        let msg = GameMessage::new(
-            MessageSource::Game("g".into()),
-            2,
-            "game:g".into(),
-            "content".into(),
-        );
-        assert!(msg.kind.is_none());
-    }
-
-    #[test]
-    fn game_message_skips_kind_when_none() {
-        let msg = GameMessage::new(
-            MessageSource::Game("g".into()),
-            1,
-            "subj".into(),
-            "c".into(),
-        );
-        let s = serde_json::to_string(&msg).expect("serialize");
-        assert!(!s.contains("\"kind\""), "kind field should be skipped: {s}");
-    }
-
-    #[test]
-    fn game_message_skips_event_fields_when_none() {
-        let msg = GameMessage::new(
-            MessageSource::Game("g".into()),
-            1,
-            "subj".into(),
-            "c".into(),
-        );
-        let s = serde_json::to_string(&msg).expect("serialize");
-        assert!(!s.contains("\"event\""), "event should be skipped: {s}");
-        assert!(
-            !s.contains("\"event_id\""),
-            "event_id should be skipped: {s}"
-        );
-    }
-
-    #[test]
-    fn game_message_with_event_populates_event_and_event_id() {
+    fn with_event_populates_event_and_event_id() {
         let event = GameEvent::TributeRest {
             tribute_id: Uuid::new_v4(),
             tribute_name: "Alice".into(),
         };
-        let msg = GameMessage::with_event(
+        let msg = with_event(
             MessageSource::Tribute("t".into()),
             3,
             "tribute:t".into(),
@@ -582,7 +404,7 @@ mod tests {
             event.clone(),
         );
         assert_eq!(
-            msg.structured_event()
+            structured_event(&msg)
                 .expect("event present")
                 .expect("decode ok"),
             event
@@ -592,12 +414,12 @@ mod tests {
     }
 
     #[test]
-    fn game_message_with_event_kind_populates_all() {
+    fn with_event_kind_populates_all() {
         let event = GameEvent::TributeRest {
             tribute_id: Uuid::new_v4(),
             tribute_name: "Alice".into(),
         };
-        let msg = GameMessage::with_event_kind(
+        let msg = with_event_kind(
             MessageSource::Tribute("t".into()),
             3,
             "tribute:t".into(),
@@ -606,7 +428,7 @@ mod tests {
             MessageKind::AllianceFormed,
         );
         assert_eq!(
-            msg.structured_event()
+            structured_event(&msg)
                 .expect("event present")
                 .expect("decode ok"),
             event
@@ -616,12 +438,12 @@ mod tests {
     }
 
     #[test]
-    fn game_message_event_roundtrip_serde() {
+    fn with_event_roundtrip_serde() {
         let event = GameEvent::TributeRest {
             tribute_id: Uuid::new_v4(),
             tribute_name: "Alice".into(),
         };
-        let msg = GameMessage::with_event(
+        let msg = with_event(
             MessageSource::Tribute("t".into()),
             3,
             "tribute:t".into(),
@@ -633,23 +455,5 @@ mod tests {
         assert_eq!(msg, back);
         assert!(back.event.is_some());
         assert!(back.event_id.is_some());
-    }
-
-    #[test]
-    fn game_message_legacy_row_without_event_fields_deserializes() {
-        // Legacy row: no `event` / `event_id` / `kind` fields present.
-        let ts = Utc.timestamp_nanos(0);
-        let raw = json!({
-            "identifier": "abc",
-            "source": { "type": "Game", "value": "g1" },
-            "game_day": 1,
-            "subject": "game:g1",
-            "timestamp": ts.timestamp_nanos_opt().unwrap(),
-            "content": "hello",
-        });
-        let msg: GameMessage = serde_json::from_value(raw).expect("deserialize legacy row");
-        assert!(msg.kind.is_none());
-        assert!(msg.event.is_none());
-        assert!(msg.event_id.is_none());
     }
 }
